cleanup and more clarity
template is done.
Created disassm screenshots to
demostrate the diff - no diff,
these are generating the exact same
release code in asm code.

we cannot run the code since the
addresses are not allowed outside the target
we canot reinterpret_cast<P*>(v) where value
is address that is like 0x25U. it is not
going to work.

But this is ok to demostrate. since we
are not going to run it on x86_64